+++ internal/gitserver/integration_tests/commits_test.go
@@ -165,7 +165,7 @@ func mustParseDate(s string, t *testing.T) *time.Time {
 }
 
 func TestHead(t *testing.T) {
-	client := gitserver.NewTestClient(http.DefaultClient, database.NewMockDB(), gitserverAddresses)
+	client := gitserver.NewTestClient(http.DefaultClient, database.NewMockDB(), GitserverAddresses)
 	t.Run("basic", func(t *testing.T) {
 		gitCommands := []string{
 			"GIT_COMMITTER_NAME=a GIT_COMMITTER_EMAIL=a@a.com GIT_COMMITTER_DATE=2006-01-02T15:04:05Z git commit --allow-empty -m foo --author='a <a@a.com>' --date 2006-01-02T15:04:05Z",
